Move Razor newline helper to compiler
