+++ src/Razor/src/Razor/test/Microsoft.CodeAnalysis.Razor.CohostingShared.UnitTests/Endpoints/CohostDocumentCompletionEndpointTest.cs
@@ -1857,7 +1857,7 @@ public partial class CohostDocumentCompletionEndpointTest(ITestOutputHelper test
         }
 
         using var _ = HashSetPool<string>.GetPooledObject(out var labelSet);
-        labelSet.UnionWith(result.Items.SelectAsArray(static item => item.Label));
+        labelSet.AddRange(result.Items.SelectAsArray((item) => item.Label));
 
 #if !VSCODE
         Assert.DoesNotContain(InvalidLabel, labelSet);
